+++ docs/src/components/side-nav/index.tsx
@@ -3,11 +3,11 @@ import { Link, navigate } from 'gatsby'
 import React, { HTMLAttributes } from 'react'
 import Theme, { useTheme } from '@pluralsight/ps-design-system-theme'
 
-interface Props extends HTMLAttributes<HTMLDivElement> {}
-
 import styles from './index.module.css'
 import logoDark from './logo-dark.png'
 import logoLight from './logo-light.png'
+
+interface Props extends HTMLAttributes<HTMLDivElement> {}
 
 export const SideNav: React.FC<Props> = () => {
   return (
@@ -15,7 +15,7 @@ export const SideNav: React.FC<Props> = () => {
       <Logo />
       <VerticalTabs>
         <VerticalTabs.Group>
-          {items.map((section, i) => {
+          {items.map(section => {
             const sectionHeader = (
               <VerticalTabs.Tier1.Header>
                 {section.header.title}