+++ packages/emptystate/src/index.js
@@ -1,3 +1,0 @@
-export { default } from './react/index.js'
-export * from './vars/index.js'
-export { default as css, sizeClasses, themeClasses } from './css/index.js'