+++ packages/position/src/js/__specs__/index.spec.js
@@ -620,12 +620,8 @@ describe('#edgeCases', () => {
 
 function randStr() {
   return (
-    Math.random()
-      .toString(36)
-      .substring(2, 15) +
-    Math.random()
-      .toString(36)
-      .substring(2, 15)
+    Math.random().toString(36).substring(2, 15) +
+    Math.random().toString(36).substring(2, 15)
   )
 }