+++ packages/util/src/__specs__/key-mirror.ts
@@ -1,0 +1,13 @@
+import { keyMirror } from '..'
+
+describe('#keyMirror', () => {
+  it('return a object of mirrored string values', () => {
+    expect(keyMirror('a', 'b', 'c')).toMatchInlineSnapshot(`
+      Object {
+        "a": "a",
+        "b": "b",
+        "c": "c",
+      }
+    `)
+  })
+})