+++ packages/util/src/types/index.ts
@@ -1,0 +1,9 @@
+import { ForwardRefExoticComponent, RefAttributes } from 'react'
+
+export type RefForwardingComponent<
+  P = Record<string, unknown>,
+  E = Element,
+  S = Record<string, unknown>
+> = ForwardRefExoticComponent<P & RefAttributes<E>> & S
+
+export type ValueOf<T> = T[keyof T]